# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Api Diliké",
        "version": "1.0"
    },
    "paths": {
        "/api/adresses": {
            "get": {
                "tags": [
                    "adresses"
                ],
                "summary": "liste des adresses",
                "description": "liste de de toutes les adresses",
                "operationId": "adresses_list",
                "responses": {
                    "200": {
                        "description": "Afficher la liste de de toutes les adresses"
                    }
                }
            },
            "post": {
                "tags": [
                    "adresses"
                ],
                "summary": "Créer une adresse",
                "description": "Créer d'une adresse",
                "operationId": "adresses",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "entreprise": {
                                        "type": "string"
                                    },
                                    "position": {
                                        "type": "string"
                                    },
                                    "ville": {
                                        "type": "string"
                                    },
                                    "quartier": {
                                        "type": "string"
                                    },
                                    "rue": {
                                        "type": "string"
                                    },
                                    "repere": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "entreprise": "ISTM",
                                    "position": "position 45 p4",
                                    "ville": "Abidjan",
                                    "quartier": "Yopougon-Koweit",
                                    "repere": "Devant la pharmacie"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/adresses/{id}": {
            "get": {
                "tags": [
                    "adresses"
                ],
                "summary": "Obtenir une seul adresse",
                "description": "Obtenir une seul adresse",
                "operationId": "one_adresse",
                "responses": {
                    "200": {
                        "description": "Obtenir une seul adresse"
                    }
                }
            },
            "put": {
                "tags": [
                    "adresses"
                ],
                "summary": "Mettre à jour une adresse",
                "description": "Mise à jour d'une adresse",
                "operationId": "adresses_update",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "entreprise": {
                                        "type": "string"
                                    },
                                    "position": {
                                        "type": "string"
                                    },
                                    "ville": {
                                        "type": "string"
                                    },
                                    "quartier": {
                                        "type": "string"
                                    },
                                    "repere": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "entreprise": "ISTM",
                                    "position": "position 45 p4",
                                    "ville": "Abidjan",
                                    "quartier": "Yopougon-Koweit",
                                    "repere": "Devant la pharmacie"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            },
            "delete": {
                "tags": [
                    "adresses"
                ],
                "summary": "Supprimer une adresse",
                "description": "Supprimer une adress",
                "operationId": "delete_adresses",
                "responses": {
                    "200": {
                        "description": "Supprimer une adresse"
                    }
                }
            }
        },
...
    }
}